+++ gnl/get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: hyobicho <hyobicho@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/05 16:26:38 by hyobicho          #+#    #+#             */
-/*   Updated: 2022/12/14 01:43:41 by hyobicho         ###   ########.fr       */
+/*   Updated: 2022/12/17 20:11:29 by hyobicho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,13 +23,13 @@
 typedef struct s_info
 {
 	char	buffer[BUFFER_SIZE];
-	int		idx;
-	int		rbyte;
-	int		len;
-	int		total;
+	ssize_t		idx;
+	ssize_t		rbyte;
+	ssize_t		len;
+	ssize_t		total;
 }	t_info;
 
-void	*ft_memcpy(void *dst, const void *src, int n);
+void	*ft_memcpy(void *dst, const void *src, ssize_t n);
 char	*ft_strjoin(char *s1, t_info *data);
 char	*get_next_line(int fd);
