+++ gnl/get_next_line_utils_bonus.c
@@ -6,17 +6,17 @@
 /*   By: hyobicho <hyobicho@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/05 16:26:18 by hyobicho          #+#    #+#             */
-/*   Updated: 2022/12/15 18:15:50 by hyobicho         ###   ########.fr       */
+/*   Updated: 2022/12/17 20:12:25 by hyobicho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "get_next_line_bonus.h"
 
-void	*ft_memcpy(void *dst, const void *src, int n)
+void	*ft_memcpy(void *dst, const void *src, ssize_t n)
 {
 	unsigned char	*dst1;
 	unsigned char	*src1;
-	int				i;
+	ssize_t				i;
 
 	if (dst == NULL && src == NULL)
 		return (dst);
